First block(final final version)

+++ src/firstBlock/MyScanner.java
@@ -4,41 +4,41 @@ import java.util.Scanner;
 
 public class MyScanner {
 
-    public static int InputNum(){
-        int number;
+    public static double InputNum(){
+        double number;
         Scanner scanner = new Scanner(System.in);
-        while (!scanner.hasNextInt()) {
+        while (!scanner.hasNextDouble()) {
             System.out.println("Only integer");
             scanner.next();
         }
-        number = scanner.nextInt();
+        number = scanner.nextDouble();
         return number;
     }
 
-    public static int InputRad(){
+    public static double InputRad(){
         Scanner scanner = new Scanner(System.in);
-        int number;
+        double number;
         do {
             System.out.println("Please enter a radius of the ring (must be positive)");
-            while (!scanner.hasNextInt()) {
+            while (!scanner.hasNextDouble()) {
                 System.out.println("That not a number!");
                 scanner.next();
             }
-            number = scanner.nextInt();
+            number = scanner.nextDouble();
         } while (number <= 0);
         return number;
     }
 
-    public static int InputWeight(){
+    public static double InputWeight(){
         Scanner scanner = new Scanner(System.in);
-        int number;
+        double number;
         do {
             System.out.println("Please enter a weight of dinosaur (must be positive)");
-            while (!scanner.hasNextInt()) {
+            while (!scanner.hasNextDouble()) {
                 System.out.println("That not a number!");
                 scanner.next();
             }
-            number = scanner.nextInt();
+            number = scanner.nextDouble();
         } while (number <= 0);
         return number;
     }
