+++ src/firstBlock/task7/task7.java
@@ -5,12 +5,12 @@ import firstBlock.MyScanner;
 public class task7 {
     public static void main(String[] args) {
         System.out.println("Input 2 integer numbers (after every num tap Enter)");
-        int a = MyScanner.InputNum();
-        int b = MyScanner.InputNum();
+        double a = MyScanner.InputNum();
+        double b = MyScanner.InputNum();
         changeNums(a,b);
     }
 
-    private static void changeNums(int a, int b){
+    private static void changeNums(double a, double b){
         a = a + b;
         b = b - a;
         b = -b;
